Add a dedicated plan for `alter reroute retry failed`.
It allows us to drop GenericDDLPlan and related to it
helper classes.
